requirements.txt
aiohttp
aioredis

args
asgiref
async-generator
async-timeout

autobahn

attrs

channels
channels-redis

cloudinary
django-cloudinary-storage
dj-database-url
Django
django-crispy-forms
django-heroku
django-summernote
docutils
django-resized
django-jazzmin
django-allauth
django-taggit
django-agora

gTTS
gTTS-token
gunicorn
hiredis
html5
hyperlink

idna
jedi
Jinja2
json5
jsonschema

Pillow

psycopg2


pyOpenSSL

requests
requests-toolbelt
social-auth-app-django
urllib3
virtualenv
virtualenv-clone
virtualenvwrapper-win

websockets
Werkzeug
whitenoise